requirements.txt
ImageHash>=4.3.0
redis>=4.3.4
Pillow>=9.2.0